Add: atomic game award

+++ labs/lab1/StatRecord.cs
@@ -3,10 +3,10 @@ namespace Lab1
   class StatRecord
   {
     public string OpponentName { get; }
-    public int Rating { get; }
+    public decimal Rating { get; }
     public bool IsWin { get; }
 
-    public StatRecord(string opponentName, int rating, bool isWin)
+    public StatRecord(string opponentName, decimal rating, bool isWin)
     {
       OpponentName = opponentName;
       Rating = rating;
